requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
fpdf==1.7.2
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.0.1
networkx==3.5
numpy==1.26.4
open_clip_torch==3.1.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.16.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.31
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tensorflow==2.19.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.3
